refactor: update api text
