+++ src/api/product-api/product-api-types.ts
@@ -1,0 +1,3 @@
+export type Product = {
+  id: number;
+};
