+++ src/api/product-api/product-api.ts
@@ -1,0 +1,5 @@
+export class ProductAPI {
+  public static async getProducts(): Promise<void> {
+    console.log('Products');
+  }
+}
