+++ src/app/app.ts
@@ -6,7 +6,7 @@ import WrapperView from './view/wrapper';
 import { Pages } from './router/pages';
 import { View } from './view/view';
 import State from './state/state';
-import { AuthAPI } from '../api/authAPI/authAPI';
+import { AuthAPI } from '../api/auth-api/auth-api';
 
 class App {
   private router: Router;
